# Introduce baseline JIT compiler