--- a/src/utils.h
+++ b/src/utils.h
@@ -3,6 +3,8 @@
 #include <sys/time.h>
 #include <time.h>
 
+#define GOLDEN_RATIO_32 0x61C88647
+
 /* Obtain the system 's notion of the current Greenwich time.
  * TODO: manipulate current time zone.
  */