… src/emulate.c
… src/riscv_private.h
… src/rv32_template.c
… tools/gen-jit-template.py