The baseline JIT compiler design involves tracing an EBB when its usage frequency exceeds a predetermined threshold. We then utilize a code generator to convert the instruction sequence into C code. The C code is subsequently compiled using mir, and the resulting target machine code is stored in the code cache for future utilization.

The primary objective of introducing the baseline JIT compiler is to enhance the execution speed of RISC-V instructions with minimal modification. This implementation requires three additional components: a code generator, mir compiler, and code cache, while maintaining the original design of the interpreter. Furthermore, this baseline JIT compiler serves as the foundational target for future improvements.

|Metric|  master| baseline JIT|Speedup|
|--------| -------- | -------- | -------- | 
| CoreMark | 1352.843 (Iterations/Sec)|1740.881 (Iterations/Sec) |+28.6％|
| dhrystone  |1146 DMIPS |2368.33DMIPS |+106%|