@jserv 2023-06-01
Resolve #136 before integrating baseline JIT compiler.