@jserv 2023-06-03
The proposed changes aim to achieve the following objectives:
1. Elimination of code duplication between `emulate.c` and `codegen.c`.
2. Reuse of the RV32 interpreter implementation for the baseline JIT.

These refinements focus on improving code organization and minimizing duplication. By reusing the existing RV32 interpreter implementation, we can expedite the development of the JIT compiler while maintaining consistency between the interpreter and JIT modes of execution.

The new branch `wip/jit` lays the groundwork for the future development of a baseline JIT compiler. By decoupling the RV32 instructions, we create reusable implementation templates that can be leveraged for code generation in the JIT compiler. This decoupling enables straightforward and flexible code generation strategies, eliminating the need for code duplication between the interpreter and MIR-based JIT compiler.

Assuming the existence of a dedicated shell script capable of generating C-string operations for each `RVOP` entry in `src/codegen.c` from `src/rv32_template.c`, the following procedure can be followed:
1. Enable the JIT compiler in the top-level configurations.
2. Generate `src/codegen.c` from `src/rv32_template.c` using the aforementioned script.
3. The file `src/emulate.c` includes the generated `src/codegen.c` for the JIT execution path.
4. The runtime determines the execution path, whether through the interpreter or JIT compilation, based on heuristic algorithms or thresholds.

This pull request should be on top of branch `wip/jit`, and you will have to rebase it and rework.