@jserv 2023-06-06
TODO: benchmarking on the following machine:
* eMag 8180 Arm server w/ GNU/Linux
* Apple M1 (or other Apple Silicon) w/ macOS

We have to ensure the benefits.