@jserv 2023-06-12
It fails to build on macOS:
```shell
$ make  
Makefile:19: *** missing separator.  Stop.
```

System information:
```
$ make --version | head -1
GNU Make 3.81
```