@qwe661234 2023-06-14
> It fails to build on macOS:
> 
> ```shell
> $ make  
> Makefile:19: *** missing separator.  Stop.
> ```
> 
> System information:
> 
> ```
> $ make --version | head -1
> GNU Make 3.81
> ```

Solved