@jserv 2023-06-14
Runtime failure on macOS. Reproduce:
```shell
$ git clone https://github.com/sysprog21/rv32emu
$ cd rv32emu
$ gh pr checkout 135
$ make ENABLE_JIT=1
$ make check
```

Error messages:
```
Assertion failed: (cache->jitcode != MAP_FAILED), function cache_create, file cache.c, line 316.
Assertion failed: (cache->jitcode != MAP_FAILED), function cache_create, file cache.c, line 316.
Assertion failed: (cache->jitcode != MAP_FAILED), function cache_create, file cache.c, line 316.
Running hello.elf ... Failed.
```